split out the writer
